extract parameterization code from execute.py into parameterize.py
